Rework how selection works
In order to allow selections to be kept during resizing
and simplify possible implementation of multi selection in the future,
rework how line selection are stored

Before this change, normalized line selected regions are stored as
two numbers and uses virtual line coordinates.
After this change selected regions are connected to each line
as an optional list of selected segments.