Move from width/mwidth of line to size/caps
The later two has an assumption that code should not
look beyond line->size. And line->caps is only used
for faster resizing.

This will faciliate moving from simple allocator
used currently to the array of pool allocators
that would store pagraphs. It would remove the
need to move line contents.

This change also requires to store screen lines
in paragraphs longer than visual lines. In this
case screen would store line views, not the raw
lines. The only hard thing in that approach is
the need to carefully split and merge paragraphs
when required. Split must happen on write to
any part of the wrapped paragraph except the last.
Merge must be performed if soft wrap is performed
during printing and the current paragraph is not
the last on screen.