Fix off-by-one in coordinate clamping
This caused the lines to be cleared improperly, when erase is performed
on the edge of the screen.

e.g. zsh uses \033[J after each completion item. And this bug led to
the improper rendering of completion suggestions on the last line of
the screen.

Reproducer:
printf "\033[1000Btree.h\033[J"; cat