Always damage screen on altscreen switch
We were relying before on screen erasing as the damage source, but
since the erasing is optimized to avoid touching empty lines, they were
not getting damaged, which caused some artifacts in corner case like
running "less" with an empty altscreen.

Damaging lines is now fast since we have a flag to damage the whole
line, so it won't bring a lot of overhead to altscreen switching.

Also make screen_swap_screen() static, since it is not used anywhere
besides the function directly below it.